Terminando el proyecto, mejorando el readme, mejorando las exceptions sobre numeros negativos, usando BigDecimal para mejorar la logica de los decimales

+++ src/main/java/com/expense_tracker/repository/ExpenseRepository.java
@@ -1,13 +1,15 @@
 package com.expense_tracker.repository;
 
 import com.expense_tracker.model.Expense;
-
+import java.math.BigDecimal;
 import java.util.List;
 
 public  interface ExpenseRepository {
+
     public Expense save(Expense expense);
     public List<Expense> findAll();
     public   List<Expense> findForMonth(int month);
-    public  double findSummary();
+    public BigDecimal findSummary();
     public boolean delete(int expenseId);
+
 }
